backend/requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
pydantic
pydantic-settings
pytest
